1.3
Added Strip after haul option
