+++ AutoStripOnHaul.csproj
@@ -47,9 +47,11 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Source\AutoStripOnHaul.cs" />
+    <Compile Include="Source\Patches\Pawn_CarryTracker_TryDropCarriedThing.cs" />
     <Compile Include="Source\Settings\Settings.cs" />
     <Compile Include="Source\Patches\Pawn_CarryTracker_TryStartCarry.cs" />
     <Compile Include="Source\Settings\SettingsData.cs" />
+    <Compile Include="Source\Utility\StripUtility.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="1.3\Assemblies\AutoStripOnHaul.dll" />
